requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
sqlalchemy_utils
pyjwt
libvirt-python
python-pam
uwsgi
werkzeug
paramiko